+++ samples/qa-videoplayer/src/views/MainPanel.js
@@ -93,7 +93,7 @@ const SelectableVideoPlayer = class extends React.Component {
 				ref={this.setVideo}
 			>
 				<MediaControls>
-					<Button className={selecting ? css.selecting : ''} onTap={this.handleToggleSelection}>repeat</Button>
+					<Button className={selecting ? css.selecting : ''} onClick={this.handleToggleSelection}>repeat</Button>
 				</MediaControls>
 				<source src="http://media.w3.org/2010/05/video/movie_300.mp4" />
 			</VideoPlayer>